+++ examples/ukrop_338lm_300gr_smk.py
@@ -35,7 +35,7 @@ zero = Shot(weapon=weapon, ammo=ammo, atmo=zero_atmo)
 zero_distance = Distance.Meter(100)
 
 config: BaseEngineConfigDict = {}
-calc = Calculator(config=config, engine="cythonized_rk4_engine")
+calc = Calculator(config=config, engine="cython+rk4")
 calc.set_weapon_zero(zero, zero_distance)
 
 current_atmo = Atmo(